+++ examples/tables.js
@@ -1,7 +1,7 @@
 /* eslint react/prop-types:0*/
 import React from 'react';
 import { render } from 'react-dom';
-import ResponsiveTable from '../src/index';
+import Table from '../src';
 import {
     SIMPLE_TABLE,
     SIMPLE_TABLE_WITH_FOOTER,
@@ -55,30 +55,30 @@ render(
         <div className="ffe-section-wrapper">
             <div className="ffe-content-container">
 
-                <h2 className="ffe-h2">Responsive table</h2>
-                <ResponsiveTable columns={SIMPLE_TABLE.columns} data={SIMPLE_TABLE.dataObject}/>
+                <h2 className="ffe-h2">Table</h2>
+                <Table columns={SIMPLE_TABLE.columns} data={SIMPLE_TABLE.dataObject}/>
 
-                <h2 className="ffe-h2">Responsive table without left padding on first column&#39;s content</h2>
-                <ResponsiveTable
+                <h2 className="ffe-h2">Table without left padding on first column&#39;s content</h2>
+                <Table
                     columns={SIMPLE_TABLE.columns}
                     data={SIMPLE_TABLE.dataObject}
                     alignLeft={true}
                 />
 
-                <h2 className="ffe-h2">Responsive table with columns hidden on tablet</h2>
-                <ResponsiveTable columns={SIMPLE_TABLE_HIDE_ON_TABLET.columns} data={SIMPLE_TABLE_HIDE_ON_TABLET.dataObject}/>
+                <h2 className="ffe-h2">Table with columns hidden on tablet</h2>
+                <Table columns={SIMPLE_TABLE_HIDE_ON_TABLET.columns} data={SIMPLE_TABLE_HIDE_ON_TABLET.dataObject}/>
 
-                <h2 className="ffe-h2">Responsive table with columns hidden on mobile</h2>
-                <ResponsiveTable columns={SIMPLE_TABLE_HIDE_ON_MOBILE.columns} data={SIMPLE_TABLE_HIDE_ON_MOBILE.dataObject}/>
+                <h2 className="ffe-h2">Table with columns hidden on mobile</h2>
+                <Table columns={SIMPLE_TABLE_HIDE_ON_MOBILE.columns} data={SIMPLE_TABLE_HIDE_ON_MOBILE.dataObject}/>
 
-                <h2 className="ffe-h2">Responsive table with 2 column layout on mobile</h2>
-                <ResponsiveTable columnLayoutMobile={true} columns={SIMPLE_TABLE_COLUMNS_ON_MOBILE.columns} data={SIMPLE_TABLE_COLUMNS_ON_MOBILE.dataObject}/>
+                <h2 className="ffe-h2">Table with 2 column layout on mobile</h2>
+                <Table columnLayoutMobile={true} columns={SIMPLE_TABLE_COLUMNS_ON_MOBILE.columns} data={SIMPLE_TABLE_COLUMNS_ON_MOBILE.dataObject}/>
 
-                <h2 className="ffe-h2">Responsive table with footer and caption</h2>
-                <ResponsiveTable columns={SIMPLE_TABLE_WITH_FOOTER.columns} data={SIMPLE_TABLE_WITH_FOOTER.dataObject} caption="This is a caption"/>
+                <h2 className="ffe-h2">Table with footer and caption</h2>
+                <Table columns={SIMPLE_TABLE_WITH_FOOTER.columns} data={SIMPLE_TABLE_WITH_FOOTER.dataObject} caption="This is a caption"/>
 
-                <h2 className="ffe-h2">Responsive table with expandable rows and column layout on mobile</h2>
-                <ResponsiveTable
+                <h2 className="ffe-h2">Table with expandable rows and column layout on mobile</h2>
+                <Table
                     columnLayoutMobile={true}
                     columns={SIMPLE_TABLE_EXPANDABLE.columns}
                     data={SIMPLE_TABLE_EXPANDABLE.dataObject}
@@ -86,10 +86,10 @@ render(
                 />
 
                 <h2 className="ffe-h2">Sortable table</h2>
-                <ResponsiveTable columns={SIMPLE_TABLE.columns} data={SIMPLE_TABLE.dataObject} sortable={true}/>
+                <Table columns={SIMPLE_TABLE.columns} data={SIMPLE_TABLE.dataObject} sortable={true}/>
 
                 <h2 className="ffe-h2">Sortable table with offset and limit (5th to 15th item) and screen-reader-only caption</h2>
-                <ResponsiveTable
+                <Table
                     columns={SIMPLE_TABLE_SORTABLE_PAGINATED.columns}
                     data={SIMPLE_TABLE_SORTABLE_PAGINATED.dataObject}
                     sortable={true}
@@ -99,9 +99,9 @@ render(
                     srOnlyCaption={true}
                 />
 
-                <h2 className="ffe-h2">Sortable, expandable, captioned, footerized, condensed, small headers, column layout on mobile</h2>
-                <ResponsiveTable
-                    columnLayoutMobile={true}
+                <h2 className="ffe-h2">Sortable, expandable, captioned, footerized, condensed, small headers, does not responsively break (will hide columns instead)</h2>
+                <Table
+                    breakpoint={ 'none' }
                     columns={SIMPLE_TABLE_SORTABLE_EXPANDABLE.columns}
                     data={SIMPLE_TABLE_SORTABLE_EXPANDABLE.dataObject}
                     expandedContentMapper={ lazyExpandedContentMapper }